fix memleak; pr feedback

+++ src/bsl/sample_pp/PolicyParser.c
@@ -187,7 +187,7 @@ static int BSLP_PolicyOptions_SC1(BSLB_VariantPtrMap_t options, const char *id_s
                 return BSL_ERR_POLICY_CONFIG;
             }
             BSL_Variant_SetInt64(opt, as_int);
-            BSL_LOG_DEBUG("     %s: %zu", id_str, as_int);
+            BSL_LOG_DEBUG("     %s: %"PRId64, id_str, as_int);
         }
     }
     else if (0 == strcmp(id_str, "scope_flags"))
@@ -201,7 +201,7 @@ static int BSLP_PolicyOptions_SC1(BSLB_VariantPtrMap_t options, const char *id_s
                 return BSL_ERR_POLICY_CONFIG;
             }
             BSL_Variant_SetInt64(opt, as_int);
-            BSL_LOG_DEBUG("     %s: %zu", id_str, as_int);
+            BSL_LOG_DEBUG("     %s: %"PRId64, id_str, as_int);
         }
     }
     else if (0 == strcmp(id_str, "key_wrap"))
@@ -215,7 +215,7 @@ static int BSLP_PolicyOptions_SC1(BSLB_VariantPtrMap_t options, const char *id_s
                 return BSL_ERR_POLICY_CONFIG;
             }
             BSL_Variant_SetInt64(opt, (int64_t)as_bool);
-            BSL_LOG_DEBUG("     %s: %zu", id_str, as_bool);
+            BSL_LOG_DEBUG("     %s: %"PRId64, id_str, (int64_t) as_bool);
         }
     }
     else
@@ -257,7 +257,7 @@ static int BSLP_PolicyOptions_SC2(BSLB_VariantPtrMap_t options, const char *id_s
                 return BSL_ERR_POLICY_CONFIG;
             }
             BSL_Variant_SetInt64(opt, as_int);
-            BSL_LOG_DEBUG("     %s: %zu", id_str, as_int);
+            BSL_LOG_DEBUG("     %s: %"PRId64, id_str, as_int);
         }
     }
     else if (0 == strcmp(id_str, "aad_scope"))
@@ -271,7 +271,7 @@ static int BSLP_PolicyOptions_SC2(BSLB_VariantPtrMap_t options, const char *id_s
                 return BSL_ERR_POLICY_CONFIG;
             }
             BSL_Variant_SetInt64(opt, as_int);
-            BSL_LOG_DEBUG("     %s: %zu", id_str, as_int);
+            BSL_LOG_DEBUG("     %s: %"PRId64, id_str, as_int);
         }
     }
     else if (0 == strcmp(id_str, "key_wrap"))
@@ -285,7 +285,7 @@ static int BSLP_PolicyOptions_SC2(BSLB_VariantPtrMap_t options, const char *id_s
                 return BSL_ERR_POLICY_CONFIG;
             }
             BSL_Variant_SetInt64(opt, (int64_t)as_bool);
-            BSL_LOG_DEBUG("     %s: %zu", id_str, as_bool);
+            BSL_LOG_DEBUG("     %s: %"PRId64, id_str, (int64_t) as_bool);
         }
     }
     else
